+++ NEWS.md
@@ -4,11 +4,12 @@
 * Package now works based on raw data with automated model estimation using lavaan (function `ipv_est`). This is now the recommended workflow!
 * New raw example data (`HEXACO`) from the open psychometrics project (\url{https://openpsychometrics.org/}) (see `?HEXACO`).
 * Create an overview over all (squared) factor loadings in all three models on which a nested chart is based (function `item_overview`).
+* Refined method available to compute center distances for groups of items (parameter `cd_method`). This is now the default method!
 * Function for convenient changing of test, facet, or item labels (function `rename`).
 * Rotate the tick label in facet charts and nested charts to avoid overlap (parameter `rotate_tick_label`).
 * Set the limit of the grid in item charts (parameter `grid_limit`).
-* Set the order of facets for all charts types (parameter: `facet_order`).
-* Set the order of tests for nested charts (parameter: `test_order`).
+* Set the order of facets for all charts types (parameter `facet_order`).
+* Set the order of tests for nested charts (parameter `test_order`).
 * Add an overall title for all chart types (parameter `title`).
 * Draw an xarrow in a nested chart that end on a test circle (instead of a facet circle) if `facet1`or `facet2` is specified as `NA` for that arrow.
 * Overhaul of the documentation in the Vignette.